[fix] 여행지 위도 경도 자료형 변경

+++ src/main/java/com/deploy/Travalue/travel/service/dto/response/TravelScheduleDto.java
@@ -13,11 +13,11 @@ public class TravelScheduleDto {
 
     private String address;
 
-    private Long latitude;
+    private Double latitude;
 
-    private Long longitude;
+    private Double longitude;
 
-    public static TravelScheduleDto of (String name, String address, Long latitude, Long longitude) {
+    public static TravelScheduleDto of (String name, String address, Double latitude, Double longitude) {
         TravelScheduleDto response = new TravelScheduleDto(
                 name,
                 address,
